86b256aag: refactored acticity details component
