+++ server/src/components/mentorStudentTaskDetailDialog.tsx
@@ -1,0 +1,87 @@
+// src/components/MentorStudentTaskDetailDialog.tsx
+
+import React from 'react';
+import { AlertDialog, AlertDialogTrigger, AlertDialogContent, AlertDialogHeader, AlertDialogTitle, AlertDialogDescription, AlertDialogFooter, AlertDialogCancel, AlertDialogAction } from '@/components/ui/alert-dialog';
+import { Input } from '@/components/ui/input'; 
+import { Textarea } from '@/components/ui/textarea'; 
+
+
+interface MentorStudentTaskDetailDialogProps {
+  taskModalOpen: boolean;
+  setTaskModalOpen: (open: boolean) => void;
+  role: string;
+  selectedUser: string;
+  studentId: string;
+  formData: { date: string };
+  workingHours: number;
+  setWorkingHours: (hours: number) => void;
+  notes: string;
+  setNotes: (notes: string) => void;
+  isEditable: boolean;
+  handleClose: () => void;
+  handleSubmit: () => void;
+}
+
+const MentorStudentTaskDetailDialog: React.FC<MentorStudentTaskDetailDialogProps> = ({
+  taskModalOpen,
+  setTaskModalOpen,
+  role,
+  selectedUser,
+  studentId,
+  formData,
+  workingHours,
+  setWorkingHours,
+  notes,
+  setNotes,
+  isEditable,
+  handleClose,
+  handleSubmit
+}) => {
+  return (
+    role === 'mentor' && selectedUser === studentId && (
+      <AlertDialog open={taskModalOpen} onOpenChange={setTaskModalOpen}>
+        <AlertDialogContent>
+          <AlertDialogHeader>
+            <AlertDialogTitle>Task Details</AlertDialogTitle>
+          </AlertDialogHeader>
+          <AlertDialogDescription>
+            <form>
+              <div className="mb-4">
+                <label>Date</label>
+                <Input type="date" value={formData.date} disabled className="text-black" />
+              </div>
+
+              <div className="mb-4">
+                <label>Working Hours</label>
+                <Input
+                  type="number"
+                  value={workingHours}
+                  onChange={(e) => setWorkingHours(Number(e.target.value) || 0)}
+                  placeholder="Enter working hours"
+                  disabled={!isEditable}
+                  className="text-black"
+                />
+              </div>
+              <div className="mb-4">
+                <label>Notes</label>
+                <Textarea
+                  value={notes}
+                  onChange={(e) => setNotes(e.target.value)}
+                  placeholder="Enter notes"
+                  disabled={!isEditable}
+                  className="text-black"
+                />
+              </div>
+            </form>
+          </AlertDialogDescription>
+          <AlertDialogFooter>
+            <AlertDialogCancel onClick={handleClose}>Cancel</AlertDialogCancel>
+            {isEditable && <AlertDialogAction onClick={handleSubmit}>Save</AlertDialogAction>}
+          </AlertDialogFooter>
+        </AlertDialogContent>
+      </AlertDialog>
+    )
+  );
+};
+
+export default MentorStudentTaskDetailDialog;
